+++ tests/integration/test_directory_watcher.py
@@ -23,6 +23,7 @@ def _get_id(item: Any, key: str) -> Any:
 def test_auto_migration(
     request_configuration: Configuration,
     storage_configuration: Configuration,
+    _common_dlm_endpoints,
 ):  # pylint: disable=too-many-locals
     """Test auto migration using directory watcher."""
     testfilename = f"group.{str(time())}"